Update project_code.sql
Top query is a detailed query joining several tables. I joined the film_category and category so I can access the genre's or the category.name, I put that under an alias as genre. The only way to access that was to go to Film -> Film_Category -> Category so that was like 2 Inner Joins but whatever.

I grouped it by inventory id firstly then I specified that I wanted the HAVING COUNT or the sales need to be more or equal to than 5.

The second query is just the summary, a lot less is going on . I'm just querying for the sale information, it's not that bad.