Refactor detailed_table
So I changed the NUMERIC from
NUMERIC(2,2) to
NUMERIC(4,2)

At most, we'll get
$xx.xx
So we only need a length of 4. Recall, the 2nd parameter for NUMERIC will specify how many decimal points you want.

I also modified the SELECT statement to cast the conversion with the data type that I mentioned above. Finally, I also decided to do a LEFT JOIN per the video on "Triggers and Procedures." He makes a good point, I wouldn't want to potentially miss any records on the LEFT table if it can relate to revenue. I'll see if my SELECT query corresponding to the summary table changes when I replace the INNER JOINS with LEFT JOINS.

Also, I set up the shell to insert data into the detailed_table and the summary_table